code/backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
flask-sqlalchemy==3.0.5
flask-marshmallow==0.15.0
flask-jwt-extended==4.5.3
flask-limiter==3.5.0
flask-bcrypt==1.0.1
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0
web3>=6.0.0
python-dotenv==1.0.0
pandas>=2.0.0
xgboost==1.7.6
joblib==1.3.2
gunicorn==21.2.0
pytest==7.4.2
pytest-cov==4.1.0
redis==5.0.0
celery==5.3.2
cryptography==41.0.4
bcrypt==4.0.1
psycopg2-binary==2.9.7
pymongo==4.5.0
sqlalchemy==2.0.21
alembic==1.12.0
requests==2.31.0
pyjwt==2.8.0
werkzeug==2.3.7
click==8.1.7
itsdangerous==2.1.2
jinja2==3.1.2
markupsafe==2.1.3
blinker==1.6.3
packaging==23.1
typing-extensions==4.8.0

